Initial revision
Initial version of an Amsterdam Schema generator.

The streaming GraphQL response builder has been subclassed
to allow the use of formatting parameters.

The formatting parameters are used to format the result
of a GraphQL query into ndjson format that adheres to
the Amsterdam Schema format.

A utility has been developed to generate Amsterdam Schema
definitions, GraphQL queries and curl commands for any
GOB catalog and collection.

It is an initial revision that has been tested for the
GOB meetbouten catalog meetbouten collection. But it
should work for any GOB catalog - collection